version 1
This version of the chatbot uses calls my web server and my web server calls a twitch API. It's set up for the glitch server, however, because glitch put's it to sleep, it shuts down, and the timers don't work correctly. New version coming soon, for hosting on other platforms.